requirements.txt
pyserial
pyserial-asyncio
cryptography
playsound
requests